Use relative paths for npm binaries
There is currently an open issue with npm related to running binaries
which are located in node_modules/.bin when the base directory contains
a %2F (/). This isn't great if we're going to adopt branch name conventions
such as feature/shiny &c as it'll cause the builds to fail in ci. Also, using
the relative path to the symlink in node_modules/.bin also fails. The only
current work-around is to provide the path to the binary that is located
with the package, hence why we no longer create the binary symlinks.