Refactor directors and add `/api` route
Refactor the proxy directors to simplify the codebase and reduce the
copy paste patterns.

Add a new `/api` route that proxies to the API router so that we can
remove the need to continually add new API routes to this florence
server, thereby reducing code complexity and the effort required to
add new functionality to florence in the future.